+++ Gruntfile.js
@@ -12,6 +12,7 @@ module.exports = function (grunt) {
     });
 
     grunt.loadNpmTasks('grunt-contrib-jshint');
+
     grunt.registerTask('default', ['jshint']);
 
     grunt.loadTasks('tasks');